# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "laterfeed",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/entries": {
      "get": {
        "tags": [
          "Feed"
        ],
        "summary": "List entries",
        "operationId": "listEntries",
        "responses": {
          "200": {
            "description": "List of entries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEntriesResponse"
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "tags": [
          "Feed"
        ],
        "summary": "Add an entry",
        "operationId": "addEntry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddEntryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Entry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/feed": {
      "get": {
        "tags": [
          "Feed"
        ],
        "summary": "Get Atom feed",
        "operationId": "getFeed",
        "responses": {
          "200": {
            "description": "Atom XML feed",
            "content": {
              "application/atom+xml": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
      "AddEntryRequest": {
        "type": "object",
        "required": [
          "url"
        ],
        "properties": {
          "source_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EntrySourceType"
              }
            ]
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "url": {
            "type": "string"
          }
        }
      },
      "EntryResponse": {
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "source_type",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "source_type": {
            "$ref": "#/components/schemas/EntrySourceType"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
      "EntrySourceType": {
        "type": "string",
        "enum": [
          "article",
          "video"
        ]
      },
      "ListEntriesResponse": {
        "type": "object",
        "required": [
          "entries"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntryResponse"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
  "tags": [
    {
      "name": "Common"
    },
    {
      "name": "Feed"
    }
  ]
}